replaced the old dashboard with the improved one
