+++ app.py
@@ -5,7 +5,6 @@ from components.layout import create_layout, register_page_callbacks
 from dash.dependencies import Output, Input
 from pages.login import register_callbacks as register_login_callbacks
 from pages.dashboard import register_callbacks as register_dashboard_callbacks
-from pages.dashboard2 import register_callbacks as register_dashboard_callbacks2
 from pages.gallery import register_callbacks as register_gallery_callbacks
 from app_data import app_data
 
@@ -18,8 +17,7 @@ app.layout = create_layout(app)
 # Registreer de callbacks
 register_page_callbacks(app)
 register_login_callbacks(app)
-# register_dashboard_callbacks(app) # TODO: uncomment this.
-register_dashboard_callbacks2(app)
+register_dashboard_callbacks(app)
 register_gallery_callbacks(app)
 
 @app.callback(
