+++ pages/dashboard.py
@@ -1,9 +1,12 @@
+# TODO: wanneer een nieuw bestand wordt geselecteerd, verwijder de oude data.
+
 import json
 from dash import html, dcc, Input, Output, State, ALL, MATCH
 import base64
 import io
 import pandas as pd
 import plotly.graph_objs as go
+import plotly.express as px
 from app_data import app_data
 from datetime import datetime
 import os
@@ -14,9 +17,11 @@ with open("data/colors.json", "r") as f:
     COLORS = json.load(f)
 
 COLOR_CODES = {}
+COLOR_NAMES = {}
 for color_name, shades in COLORS.items():
     for shade, code in shades.items():
         COLOR_CODES[f"{color_name} {shade}"] = code
+        COLOR_NAMES[code] = f"{color_name} {shade}"
 
 COLOR_EMOJIS = {
     "Lintblauw": "🟦",
@@ -29,6 +34,8 @@ COLOR_EMOJIS = {
     "Grijs": "⬜",
     # Voeg meer toe indien gewenst
 }
+
+HEATMAP_COLORS = ["Viridis", "Plasma", "Inferno", "Magma", "Cividis", "Blues", "Greens", "YlGnBu", "YlOrRd", "RdBu", "Picnic", "Jet"]
 
 def layout(app):
     return html.Div([
@@ -70,7 +77,7 @@ def layout(app):
 
         # Upload bouwblok.
         html.Div([
-            html.H4("X. Upload json-export-bestanden", style={"marginBottom": "10px"}),
+            html.H4("X. Upload json-export-bestanden fdsfdsfd", style={"marginBottom": "10px"}),
             dcc.Upload(
                 id='upload-export-data',
                 children=html.Button('📤 Upload json-export-bestanden', style={
@@ -95,64 +102,111 @@ def layout(app):
             "marginRight": "auto"
         }),
 
-        # Instellingen card
         html.Div([
+            html.H4("X. Kies het visualisatietype", style={"marginBottom": "10px"}),
+            html.Div([
+                html.Label("Chart type:", style={
+                    "fontWeight": "bold",
+                    "minWidth": "140px",
+                    "marginRight": "12px"
+                }),
+                dcc.Dropdown(
+                    id='chart-type',
+                    options=[
+                        {'label': 'Lijn (Line)', 'value': 'line'},
+                        {'label': 'Staaf (Bar)', 'value': 'bar'},
+                        {'label': 'Punten (Scatter)', 'value': 'scatter'},
+                        {'label': 'Taartdiagram (Pie)', 'value': 'pie'},
+                        {'label': 'Histogram', 'value': 'histogram'},
+                        {'label': 'Boxplot', 'value': 'box'},
+                        # {'label': 'Violinplot', 'value': 'violin'},
+                        {'label': 'Heatmap', 'value': 'heatmap'},
+                        # {'label': 'Area', 'value': 'area'},
+                        # {'label': 'Bubble', 'value': 'bubble'},
+                        # {'label': 'Funnel', 'value': 'funnel'},
+                        # {'label': 'Sunburst', 'value': 'sunburst'},
+                        # {'label': 'Treemap', 'value': 'treemap'},
+                        # {'label': 'Polar', 'value': 'polar'},
+                        # {'label': '3D Scatter', 'value': 'scatter3d'},
+                        # {'label': '3D Surface', 'value': 'surface'},
+                    ],
+                    style={"width": "100%"}
+                )
+            ], style={"display": "flex", "alignItems": "center", "marginBottom": "12px"}),
+        ], style={
+            "backgroundColor": "#f7f7f7",
+            "padding": "24px",
+            "borderRadius": "10px",
+            "boxShadow": "0 2px 8px rgba(0,0,0,0.07)",
+            "marginBottom": "30px",
+            "maxWidth": "600px",
+            "marginLeft": "auto",
+            "marginRight": "auto"
+        }),
+
+        # Instellingen card
+        html.Div(children=[
             html.H4("2. Instellingen", style={"marginBottom": "18px"}),
+
             html.Div([
                 html.Div([
-                    html.Label("X-as kolom:", style={
+                    html.Label("x-as:", style={
                         "fontWeight": "bold",
                         "minWidth": "140px",
                         "marginRight": "12px"
                     }),
                     dcc.Dropdown(id='x-axis', style={"width": "100%"})
                 ], style={"display": "flex", "alignItems": "center", "marginBottom": "12px"}),
-
-                html.Div([
-                    html.Label("Y-as kolommen:", style={
-                        "fontWeight": "bold",
-                        "minWidth": "140px",
-                        "marginRight": "12px"
-                    }),
-                    dcc.Dropdown(id='y-axis', multi=True, style={"width": "100%"})
-                ], style={"display": "flex", "alignItems": "center", "marginBottom": "12px"}),
             ], style={"marginBottom": "18px"}),
-
-            html.Label("Kleuren per Y-as:", style={"fontWeight": "bold", "marginTop": "10px"}),
-            html.Div(id='color-selectors', style={"marginBottom": "18px"}),
 
             html.Div([
                 html.Div([
-                    html.Label("Chart type:", style={
+                    html.Label("y-as:", style={
                         "fontWeight": "bold",
                         "minWidth": "140px",
                         "marginRight": "12px"
                     }),
-                    dcc.Dropdown(
-                        id='chart-type',
-                        options=[
-                            {'label': 'Lijn (Line)', 'value': 'line'},
-                            {'label': 'Staaf (Bar)', 'value': 'bar'},
-                            {'label': 'Punten (Scatter)', 'value': 'scatter'},
-                            {'label': 'Taartdiagram (Pie)', 'value': 'pie'},
-                            {'label': 'Histogram', 'value': 'histogram'},
-                            {'label': 'Boxplot', 'value': 'box'},
-                            {'label': 'Violinplot', 'value': 'violin'},
-                            {'label': 'Heatmap', 'value': 'heatmap'},
-                            {'label': 'Area', 'value': 'area'},
-                            {'label': 'Bubble', 'value': 'bubble'},
-                            {'label': 'Funnel', 'value': 'funnel'},
-                            {'label': 'Sunburst', 'value': 'sunburst'},
-                            {'label': 'Treemap', 'value': 'treemap'},
-                            {'label': 'Polar', 'value': 'polar'},
-                            {'label': '3D Scatter', 'value': 'scatter3d'},
-                            {'label': '3D Surface', 'value': 'surface'},
-                        ],
-                        value='line',
-                        style={"width": "100%"}
-                    )
+                    dcc.Dropdown(id='y-axis', style={"width": "100%"})
                 ], style={"display": "flex", "alignItems": "center", "marginBottom": "12px"}),
+            ], style={"marginBottom": "18px"}),
 
+            html.Div(id="nbins-section", children=[
+                html.Div([
+                    html.Label("nbins:", style={
+                        "fontWeight": "bold",
+                        "minWidth": "140px",
+                        "marginRight": "12px"
+                    }),
+                    dcc.Input(id='nbins', type="number", style={"width": "75%", "padding": "6px 12px", "border": "1px solid #ced4da", "borderRadius": "4px", "backgroundColor": "white", "fontSize": "14px", "height": "36px", "boxSizing": "border-box"})
+                ], style={"display": "flex", "alignItems": "space-between", "marginBottom": "12px"}),
+            ], style={"marginBottom": "18px"}),
+
+            html.Div(id="heatmap-value-section", children=[
+                html.Div([
+                    html.Label("waardekolom:", style={
+                        "fontWeight": "bold",
+                        "minWidth": "140px",
+                        "marginRight": "12px"
+                    }),
+                    dcc.Dropdown(id='heatmap-dropdown', style={"width": "100%"})
+                ], style={"display": "flex", "alignItems": "space-between", "marginBottom": "12px"}),
+            ], style={"marginBottom": "18px"}),
+
+            html.Div([
+                html.Div([
+                    html.Label("Kleur toepassen:", style={
+                        "fontWeight": "bold",
+                        "minWidth": "140px",
+                        "marginRight": "12px"
+                    }),
+                    dcc.Dropdown(id='column-to-color', style={"width": "100%"})
+                ], style={"display": "flex", "alignItems": "center", "marginBottom": "12px"}),
+            ], style={"marginBottom": "18px"}),
+
+            html.Label("Kleuren:", style={"fontWeight": "bold", "marginTop": "10px"}),
+            html.Div(id='color-selectors', style={"marginBottom": "18px"}),
+
+            html.Div([
                 html.Div([
                     html.Label("Titel van de visualisatie:", style={"fontWeight": "bold"}),
                     dcc.Input(
@@ -231,14 +285,16 @@ def register_callbacks(app):
     @app.callback(
         [Output('upload-output', 'children'),
          Output('x-axis', 'options'),
-         Output('y-axis', 'options')],
+         Output('y-axis', 'options'),
+         Output('column-to-color', 'options'),
+         Output('heatmap-dropdown', 'options', allow_duplicate=True)],
         [Input('upload-data', 'contents')],
-        [State('upload-data', 'filename')]
+        [State('upload-data', 'filename')],
+        prevent_initial_call=True
     )
     def handle_upload(contents_list, filenames):
         if contents_list is None:
-            return "❌ Geen bestanden geüpload.", [], []
-
+            return "❌ Geen bestanden geüpload.", [], [], []
         try:
             dfs = []  # Lijst om dataframes op te slaan
             for contents, filename in zip(contents_list, filenames):
@@ -260,35 +316,49 @@ def register_callbacks(app):
                 dfs.append(df)  # Voeg het dataframe toe aan de lijst
 
             if not dfs:
-                return "❌ Geen geldige CSV of Excel bestanden geüpload.", [], []
+                return "❌ Geen geldige CSV of Excel bestanden geüpload.", [], [], []
 
             # Combineer alle dataframes
-            combined_df = pd.concat(dfs, ignore_index=True)
+            combined_df = pd.concat(dfs, ignore_index=True)#.head(5) # TODO: remove .head call.
 
             # Sla het gecombineerde dataframe op in app_data
             app_data['df'] = combined_df
+            app_data['parameters'] = {
+                "dataframe": combined_df.to_dict(),
+                "length": combined_df.shape[0],
+                "columns": list(combined_df.columns),
+            }
+
+            dropdown_values = HEATMAP_COLORS
 
             # Genereer opties voor de dropdowns
             options = [{'label': col, 'value': col} for col in combined_df.columns]
-            return f"✅ {len(dfs)} bestand(en) succesvol geüpload en gecombineerd.", options, options
+            return f"✅ {len(dfs)} bestand(en) succesvol geüpload en gecombineerd.", options, options, dropdown_values, options
 
         except Exception as e:
             # Foutafhandeling
-            return f"❌ Fout bij het verwerken van de bestanden: {e}", [], []
+            return f"❌ Fout bij het verwerken van de bestanden: {e}", [], [], [], []
 
     @app.callback(
         [Output('upload-export-output', 'children'),
          Output('plot', 'figure', allow_duplicate=True),
+         Output('column-to-color', 'options', allow_duplicate=True),
          Output('x-axis', 'options', allow_duplicate=True),
          Output('y-axis', 'options', allow_duplicate=True),
          Output('chart-type', 'value', allow_duplicate=True)],
+         Output('x-axis', 'value'),
+         Output('y-axis', 'value'),
+         Output('column-to-color', 'value'),
+         Output('nbins', 'value'),
+         Output('heatmap-dropdown', 'options'),
+         Output('heatmap-dropdown', 'value'),
         [Input('upload-export-data', 'contents')],
         [State('upload-export-data', 'filename')],
         prevent_initial_call=True
     )
     def handle_export_file_upload(content, filename):
         if not filename.lower().endswith('.json'):
-            return ["❌ Upload een JSON bestand.", None, [], [], None]
+            return ["❌ Upload een JSON bestand.", None, [], [], [], None, None, None, None, 0, None, ""]
 
         _content_type, content_string = content.split(',')
         decoded = base64.b64decode(content_string)
@@ -303,21 +373,97 @@ def register_callbacks(app):
 
         graph = content["figureContents"]
         app_data["graph"] = graph
+        app_data["parameters"] = content["parameters"]
+        x = app_data["parameters"]["x"]
+        y = app_data["parameters"].get("y")
+        column_to_color = app_data["parameters"].get("columnToColor")
+        app_data['df'] = pd.DataFrame(app_data["parameters"]["dataframe"])
 
-        options = [{'label': col, 'value': col} for col in content["portalData"]["columns"]]
+        nbins = 0
+        if content["parameters"]["chartType"] == "histogram":
+            nbins = app_data["parameters"]["nbins"]
 
-        return [f"Geselecteerd bestand: {filename}", graph, options, options, content["portalData"]["chartType"]]
+        options = [{'label': col, 'value': col} for col in content["parameters"]["columns"]]
+
+        dropdown_values = []
+        value_column = None
+        if content["parameters"]["chartType"] == "heatmap":
+            dropdown_values = HEATMAP_COLORS
+            column_to_color = content["parameters"]["colorContinuousScale"]
+            value_column = content["parameters"]["valueColumn"]
+        else:
+            dropdown_values = options
+
+        return [f"✅ Geselecteerd bestand: {filename}", graph, dropdown_values, options, options, content["parameters"]["chartType"], x, y, column_to_color, nbins, options, value_column]
+
+
+    @app.callback(
+        Output('column-to-color', 'disabled'),
+        Input('chart-type', 'value'),
+        prevent_initial_call=True
+    )
+    def disable_color_dropdown(chart_type):
+        return chart_type == "scatter" or chart_type == "histogram"
+
+    @app.callback(
+        Output('y-axis', 'disabled'),
+        Input('chart-type', 'value'),
+        prevent_initial_call=True
+    )
+    def disable_y_dropdown(chart_type):
+        return chart_type == "histogram"
+
+    @app.callback(
+        Output('nbins-section', 'style'),
+        Input('chart-type', 'value'),
+    )
+    def show_nbins_section(chart_type):
+        if chart_type == "histogram":
+            return {'display': 'block'}
+        else:
+            return {'display': 'none'}
+
+    @app.callback(
+        Output('heatmap-value-section', 'style'),
+        Input('chart-type', 'value'),
+    )
+    def show_heatmap_value_section(chart_type):
+        if chart_type == "heatmap":
+            return {'display': 'block'}
+        else:
+            return {'display': 'none'}
 
     @app.callback(
         Output('color-selectors', 'children'),
-        Input('y-axis', 'value')
+        Input('column-to-color', 'value'),
+        Input('chart-type', 'value'),
     )
-    def update_color_selectors(y_columns):
-        if not y_columns:
-            return "❌ Selecteer eerst Y-as kolommen."
+    def update_color_selectors(column_to_color, chart_type):
+        if not column_to_color and not chart_type == "scatter" and not chart_type == "histogram":
+            return "❌ Selecteer de kolom die gekleurd moet worden."
         
+        if "parameters" not in app_data or "df" not in app_data:
+            return "❌ Selecteer een dataset."
+
+        if chart_type == "heatmap":
+            return "Gebruik de waardekolom dropdown voor een kleur."
+
         dropdowns = []
-        for col in y_columns:
+
+        length = 0
+        if chart_type == "line" or chart_type == "box":
+            length = app_data['df'][column_to_color].unique().shape[0]
+        elif chart_type == "scatter" or chart_type == "histogram":
+            length = 1
+        else:
+            length = app_data["parameters"]["length"]
+
+        for col in range(0, length):
+            color_from_module = None
+            if "colors" in app_data["parameters"]:
+                color_amount = len(app_data["parameters"]["colors"])
+                color_from_module = app_data["parameters"]["colors"][col % color_amount]
+
             options = []
             for color_name, shades in COLORS.items():
                 for shade, code in shades.items():
@@ -331,7 +477,7 @@ def register_callbacks(app):
             default_color = COLOR_CODES.get(default_value, "#000000")
             dropdowns.append(
                 html.Div([
-                    html.Label(f"Kleur voor {col}:", style={
+                    html.Label(f"Kleur voor {col + 1}:", style={
                         "minWidth": "120px",
                         "marginRight": "16px",
                         "fontWeight": "bold"
@@ -352,7 +498,7 @@ def register_callbacks(app):
                         dcc.Dropdown(
                             id={'type': 'color-dropdown', 'index': col},
                             options=options,
-                            value=default_value,
+                            value=COLOR_NAMES[color_from_module] if color_from_module else default_value,
                             style={
                                 'width': '200px',
                                 'display': 'inline-block',
@@ -377,74 +523,106 @@ def register_callbacks(app):
         Input('generate-btn', 'n_clicks'),
         State('x-axis', 'value'),
         State('y-axis', 'value'),
+        State('column-to-color', 'value'),
         State('chart-type', 'value'),
-        State({'type': 'color-dropdown', 'index': ALL}, 'value')  # Haal kleuren op
+        State({'type': 'color-dropdown', 'index': ALL}, 'value'), # Haal kleuren op
+        State('nbins', 'value'),
+        State('heatmap-dropdown', 'value'),
     )
-    def generate_chart(n, x, ys, chart_type, colors):
+    def generate_chart(n, x, y, column_to_color, chart_type, colors, nbins, value_column):
         df = app_data.get('df')
         fig = go.Figure()
-        if df is None or x is None or not ys:
-            return fig
 
-        for i, y in enumerate(ys):
-            naam_tint = colors[i] if i < len(colors) else None
-            color = COLOR_CODES.get(naam_tint, "#000000")
-            if chart_type == 'line':
-                fig.add_trace(go.Scatter(x=df[x], y=df[y], mode='lines+markers', name=y, line=dict(color=color)))
-            elif chart_type == 'bar':
-                fig.add_trace(go.Bar(x=df[x], y=df[y], name=y, marker=dict(color=color)))
-            elif chart_type == 'scatter':
-                fig.add_trace(go.Scatter(x=df[x], y=df[y], mode='markers', name=y, marker=dict(color=color)))
-            elif chart_type == 'area':
-                fig.add_trace(go.Scatter(x=df[x], y=df[y], fill='tozeroy', mode='lines', name=y, line=dict(color=color)))
-            elif chart_type == 'bubble':
-                fig.add_trace(go.Scatter(
-                    x=df[x], y=df[y], mode='markers', name=y,
-                    marker=dict(size=df[y], color=color, sizemode='area', sizeref=2.*max(df[y])/(40.**2), sizemin=4)
-                ))
-            elif chart_type == 'histogram':
-                fig.add_trace(go.Histogram(x=df[y], name=y, marker=dict(color=color)))
-            elif chart_type == 'box':
-                fig.add_trace(go.Box(y=df[y], name=y, marker=dict(color=color)))
-            elif chart_type == 'violin':
-                fig.add_trace(go.Violin(y=df[y], name=y, line=dict(color=color)))
-            elif chart_type == 'heatmap':
-                if len(ys) > 1 and x in df.columns:
-                    fig.add_trace(go.Heatmap(z=df[ys].values, x=df[x], y=ys, colorscale='Viridis'))
-                    break
-                else:
-                    fig.add_trace(go.Heatmap(z=[df[y].values], x=df[x], y=[y], colorscale='Viridis'))
-            elif chart_type == 'pie':
-                fig = go.Figure(go.Pie(labels=df[x], values=df[y], marker=dict(colors=[color])))
-                break
-            elif chart_type == 'donut':
-                fig = go.Figure(go.Pie(labels=df[x], values=df[y], hole=0.4, marker=dict(colors=[color])))
-                break
-            elif chart_type == 'funnel':
-                fig = go.Figure(go.Funnel(y=df[x], x=df[y], marker=dict(color=color)))
-                break
-            elif chart_type == 'sunburst':
-                fig = go.Figure(go.Sunburst(labels=df[x], parents=[""]*len(df[x]), values=df[y]))
-                break
-            elif chart_type == 'treemap':
-                fig = go.Figure(go.Treemap(labels=df[x], parents=[""]*len(df[x]), values=df[y]))
-                break
-            elif chart_type == 'polar':
-                fig.add_trace(go.Scatterpolar(r=df[y], theta=df[x], mode='lines+markers', name=y, line=dict(color=color)))
-            elif chart_type == 'scatter3d':
-                if len(ys) >= 3:
-                    fig = go.Figure(go.Scatter3d(
-                        x=df[ys[0]], y=df[ys[1]], z=df[ys[2]],
-                        mode='markers', marker=dict(color=color), name="3D Scatter"
-                    ))
-                    break
-            elif chart_type == 'surface':
-                if len(ys) >= 3:
-                    fig = go.Figure(go.Surface(
-                        z=df[ys[2]].values.reshape((len(df[ys[0]].unique()), len(df[ys[1]].unique()))),
-                        x=df[ys[0]].unique(), y=df[ys[1]].unique(), colorscale='Viridis'
-                    ))
-                    break
+        if "parameters" in app_data and "dataframe" in app_data["parameters"]:
+            df = pd.DataFrame(app_data["parameters"]["dataframe"])
+
+        select_colors = []
+        if colors is not None:
+            for color in colors:
+                color_code = COLOR_CODES.get(color)
+                if color_code is None:
+                    continue
+                select_colors.append(color_code)
+
+        if chart_type == 'bar':
+            fig = px.bar(df, x=x, y=y, color=column_to_color, color_discrete_sequence=select_colors)
+        elif chart_type == "pie":
+            fig = px.pie(df, values=x, names=y, color=column_to_color, color_discrete_sequence=select_colors)
+        elif chart_type == "line":
+            fig = px.line(df, x=x, y=y, color=column_to_color, color_discrete_sequence=select_colors)
+        elif chart_type == "scatter":
+            fig = px.scatter(df, x=x, y=y, color_discrete_sequence=select_colors)
+        elif chart_type == "histogram":
+            fig = px.histogram(df, x=x, y=y, nbins=nbins, color_discrete_sequence=select_colors)
+        elif chart_type == "box":
+            fig = px.box(df, x=x, y=y, color=column_to_color, color_discrete_sequence=select_colors)
+        elif chart_type == "heatmap":
+            heatmap_data = df.pivot(index=x, columns=y, values=value_column)
+            fig = px.imshow(
+                heatmap_data,
+                labels=dict(x=x, y=y, color=value_column),
+                color_continuous_scale=column_to_color,
+            )
+
+        # for i, y in enumerate(ys):
+        #     naam_tint = colors[i] if i < len(colors) else None
+        #     color = COLOR_CODES.get(naam_tint, "#000000")
+        #     if chart_type == 'line':
+        #         fig.add_trace(go.Scatter(x=df[x], y=df[y], mode='lines+markers', name=y, line=dict(color=color)))
+        #     elif chart_type == 'bar':
+        #         fig.add_trace(go.Bar(x=df[x], y=df[y], name=y, marker=dict(color=color)))
+        #     elif chart_type == 'scatter':
+        #         fig.add_trace(go.Scatter(x=df[x], y=df[y], mode='markers', name=y, marker=dict(color=color)))
+        #     elif chart_type == 'area':
+        #         fig.add_trace(go.Scatter(x=df[x], y=df[y], fill='tozeroy', mode='lines', name=y, line=dict(color=color)))
+        #     elif chart_type == 'bubble':
+        #         fig.add_trace(go.Scatter(
+        #             x=df[x], y=df[y], mode='markers', name=y,
+        #             marker=dict(size=df[y], color=color, sizemode='area', sizeref=2.*max(df[y])/(40.**2), sizemin=4)
+        #         ))
+        #     elif chart_type == 'histogram':
+        #         fig.add_trace(go.Histogram(x=df[y], name=y, marker=dict(color=color)))
+        #     elif chart_type == 'box':
+        #         fig.add_trace(go.Box(y=df[y], name=y, marker=dict(color=color)))
+        #     elif chart_type == 'violin':
+        #         fig.add_trace(go.Violin(y=df[y], name=y, line=dict(color=color)))
+        #     elif chart_type == 'heatmap':
+        #         if len(ys) > 1 and x in df.columns:
+        #             fig.add_trace(go.Heatmap(z=df[ys].values, x=df[x], y=ys, colorscale='Viridis'))
+        #             break
+        #         else:
+        #             fig.add_trace(go.Heatmap(z=[df[y].values], x=df[x], y=[y], colorscale='Viridis'))
+        #     elif chart_type == 'pie':
+        #         fig = go.Figure(go.Pie(labels=df[x], values=df[y], marker=dict(colors=[color])))
+        #         break
+        #     elif chart_type == 'donut':
+        #         fig = go.Figure(go.Pie(labels=df[x], values=df[y], hole=0.4, marker=dict(colors=[color])))
+        #         break
+        #     elif chart_type == 'funnel':
+        #         fig = go.Figure(go.Funnel(y=df[x], x=df[y], marker=dict(color=color)))
+        #         break
+        #     elif chart_type == 'sunburst':
+        #         fig = go.Figure(go.Sunburst(labels=df[x], parents=[""]*len(df[x]), values=df[y]))
+        #         break
+        #     elif chart_type == 'treemap':
+        #         fig = go.Figure(go.Treemap(labels=df[x], parents=[""]*len(df[x]), values=df[y]))
+        #         break
+        #     elif chart_type == 'polar':
+        #         fig.add_trace(go.Scatterpolar(r=df[y], theta=df[x], mode='lines+markers', name=y, line=dict(color=color)))
+        #     elif chart_type == 'scatter3d':
+        #         if len(ys) >= 3:
+        #             fig = go.Figure(go.Scatter3d(
+        #                 x=df[ys[0]], y=df[ys[1]], z=df[ys[2]],
+        #                 mode='markers', marker=dict(color=color), name="3D Scatter"
+        #             ))
+        #             break
+        #     elif chart_type == 'surface':
+        #         if len(ys) >= 3:
+        #             fig = go.Figure(go.Surface(
+        #                 z=df[ys[2]].values.reshape((len(df[ys[0]].unique()), len(df[ys[1]].unique()))),
+        #                 x=df[ys[0]].unique(), y=df[ys[1]].unique(), colorscale='Viridis'
+        #             ))
+        #             break
 
         app_data['graph'] = fig
         return fig
@@ -453,9 +631,10 @@ def register_callbacks(app):
         Output('publish-status', 'children'),
         Input('publish-btn', 'n_clicks'),
         State('dashboard-title', 'value'),
-        State('description', 'value')
+        State('description', 'value'),
+        State('chart-type', 'value'),
     )
-    def publish(n, title, desc):
+    def publish(n, title, desc, chart_type):
         if not n:
             return "❌ Geen actie uitgevoerd. Klik op de knop om te publiceren."
 
@@ -470,6 +649,9 @@ def register_callbacks(app):
         # Controleer of er een beschrijving is
         if not desc:
             return "❌ Beschrijving is verplicht om te publiceren."
+
+        if not chart_type:
+            return "❌ Kiez een visualisatietype."
 
         try:
             # Maak een record aan voor de grafiek
